# docs/api-swagger.yaml
openapi: 3.0.3
info:
  title: Go Data Sync HTTP API
  description: |
    # Real-Time MongoDB Data Synchronization API

    This API provides comprehensive endpoints for managing real-time MongoDB data synchronization between cloud and VM environments. The system supports OAuth2 authentication, adaptive performance optimization, and real-time monitoring.

    ## Key Features
    - **OAuth2 Client Credentials Flow** for secure authentication
    - **Real-time Data Synchronization** using MongoDB change streams
    - **Adaptive Performance Control** with automatic optimization
    - **Comprehensive Monitoring** with metrics and health endpoints
    - **WebSocket Support** for real-time updates
    - **TCP and HTTP Transport** protocols

    ## Authentication
    Most endpoints require OAuth2 Bearer token authentication. Use the `/api/auth/token` endpoint to obtain access tokens.

    ## Rate Limiting
    - Global: 100 requests/second (burst: 200)
    - Per-client: 20 requests/second (burst: 40)
    - Per-IP connections: 10/minute

  version: "2.0.0"
  contact:
    name: Go Data Sync API Support
    url: https://github.com/your-org/go-data-sync-http
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.yourdomain.com
    description: Production server
security:
  - BearerAuth: []
  - AdminAPIKey: []
paths:
...
  /api/auth/token:
    post:
      tags:
        - OAuth2 Authentication
      summary: Obtain OAuth2 Access Token
      description: |
        Implements OAuth2 Client Credentials Grant flow to obtain access tokens for API authentication.
        
        **Flow:**
        1. Client provides client_id and client_secret
        2. Server validates credentials against MongoDB
        3. Returns JWT bearer token with 1-hour expiry
        
        **Security:**
        - Client secrets are bcrypt-hashed in database
        - JWT tokens include standard RFC 7519 claims + custom claims
        - Tokens are signed with HMAC-SHA256
      operationId: getOAuth2Token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
            example:
              grant_type: "client_credentials"
              client_id: "vm_sync_3e67555ad3836e65"
              client_secret: "4684a92c64ad93d075018f0152fc5d0ac2f9d73ba2589b32c30e664e82e21e95"
      responses:
        '200':
          description: Successfully obtained access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              example:
                access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                token_type: "Bearer"
                expires_in: 3599
                scope: "[sync data:read data:write]"
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2Error'
              example:
                error: "invalid_request"
                error_description: "grant_type, client_id, and client_secret are required"
        '401':
          description: Invalid client credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2Error'
              example:
                error: "invalid_client"
                error_description: "Invalid client credentials"
  /api/auth/validate:
    post:
      tags:
        - OAuth2 Authentication
      summary: Validate JWT Token
      description: |
        Internal endpoint for validating JWT tokens and retrieving claims information.
        Used by other services to verify token authenticity and extract user context.
      operationId: validateToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: JWT token to validate
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  client_id:
                    type: string
                    example: "vm_sync_3e67555ad3836e65"
                  client_type:
                    type: string
                    example: "vm-sync"
                  app_id:
                    type: string
                    example: "vm-sync-test-working"
                  scopes:
                    type: array
                    items:
                      type: string
                    example: ["sync", "data:read", "data:write"]
                  expires_at:
                    type: integer
                    format: int64
                    description: Unix timestamp
                    example: 1758890731
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2Error'

  # ==================== OAUTH2 ADMIN ENDPOINTS ====================
  /api/auth/admin/clients:
    post:
      tags:
        - OAuth2 Administration
      summary: Create New OAuth2 Client
      description: |
        Creates new OAuth2 client credentials with cryptographically secure client_id and client_secret.
        
        **Generated Credentials:**
        - Client ID: 16 random bytes, hex-encoded with 'vm_sync_' prefix
        - Client Secret: 32 random bytes, hex-encoded (64 characters)
        - Secret stored as bcrypt hash in database
        
        **Requires Admin API Key:** `X-Admin-API-Key: admin-api-key-placeholder`
      operationId: createOAuth2Client
      security:
        - AdminAPIKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
            example:
              app_id: "vm-sync-production"
              name: "Production VM-Sync Instance"
              description: "OAuth2 client for production VM synchronization"
              scopes: ["sync", "data:read", "data:write", "metrics:read"]
              vm_metadata:
                hostname: "prod-vm-01"
                max_collections: 50
      responses:
        '200':
          description: Client created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateClientResponse'
              example:
                client_id: "vm_sync_a1b2c3d4e5f6a7b8"
                client_secret: "9f8e7d6c5b4a39284756adf123456789abcdef..."
                app_id: "vm-sync-production"
                name: "Production VM-Sync Instance"
                scopes: ["sync", "data:read", "data:write", "metrics:read"]
                created_at: "2025-09-26T12:00:00Z"
        '400':
          description: Invalid request parameters
        '401':
          description: Admin API key required
        '403':
          description: Invalid admin API key

    get:
      tags:
        - OAuth2 Administration
      summary: List All OAuth2 Clients
      description: |
        Returns list of all OAuth2 clients with sensitive information redacted.
        Client secrets are shown as "[REDACTED]" for security.
      operationId: listOAuth2Clients
      security:
        - AdminAPIKey: []
      responses:
        '200':
          description: List of OAuth2 clients
          content:
            application/json:
              schema:
                type: object
                properties:
                  clients:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClientCredentialsResponse'
                  total:
                    type: integer
                    example: 3
        '401':
          description: Admin API key required
        '403':
          description: Invalid admin API key
  /api/auth/admin/clients/{client_id}:
    delete:
      tags:
        - OAuth2 Administration
      summary: Revoke OAuth2 Client
      description: |
        Deactivates OAuth2 client credentials. The client will no longer be able to obtain new tokens.
        Existing tokens remain valid until expiry.
      operationId: revokeOAuth2Client
      security:
        - AdminAPIKey: []
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
            example: "vm_sync_a1b2c3d4e5f6a7b8"
          description: OAuth2 client ID to revoke
      responses:
        '204':
          description: Client revoked successfully
        '400':
          description: client_id is required
        '401':
          description: Admin API key required
        '403':
          description: Invalid admin API key
        '500':
          description: Failed to revoke client

  # ==================== DATA SYNCHRONIZATION ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OAuth2 JWT Bearer token authentication. Obtain tokens via `/api/auth/token` endpoint.
        
        **Usage:** `Authorization: Bearer <token>`
    
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: |
        Admin API key for administrative operations. Set via environment variable.
        
        **Usage:** `X-Admin-Key: <admin_key>`
    
    LicenseAuth:
      type: apiKey
      in: header
      name: X-License-Key
      description: |
        License key authentication for VM-sync clients.
        
        **Usage:** `X-License-Key: <license_key>`
  schemas:
...
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          enum: [Bearer]
          description: Token type (always "Bearer")
        expires_in:
          type: integer
          description: Token lifetime in seconds
          example: 3600
        scope:
          type: string
          description: Granted scopes
          example: "data_sync monitoring"
...
    CreateClientRequest:
      type: object
      required:
        - name
        - client_type
      properties:
        name:
          type: string
          description: Human-readable client name
          example: "VM Sync Client - Production"
        description:
          type: string
          description: Client description
          example: "Production VM sync client for data center 1"
        client_type:
          type: string
          enum: [vm_sync, dashboard, api_client]
          description: Type of client
        app_id:
          type: string
          description: Application identifier
          example: "vm-datacenter-1"
        scopes:
          type: array
          items:
            type: string
          description: Requested scopes
          example: ["data_sync", "monitoring"]
        expires_at:
          type: string
          format: date-time
          description: Client expiration time
          example: "2024-12-31T23:59:59Z"
    CreateClientResponse:
      type: object
      properties:
        client_id:
          type: string
          description: Generated client ID
          example: "vm_sync_3e67555ad3836e65"
        client_secret:
          type: string
          description: Generated client secret
          example: "4684a92c64ad93d075018f0152fc5d0ac2f9d73ba2589b32c30e664e82e21e95"
        name:
          type: string
          description: Client name
        client_type:
          type: string
          description: Client type
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        expires_at:
          type: string
          format: date-time
          description: Expiration timestamp
...
    OAuth2Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - server_error
          description: OAuth2 error code
        error_description:
          type: string
          description: Human-readable error description
        error_uri:
          type: string
          format: uri
          description: URL to error documentation
    
    # ==================== DATA SYNC SCHEMAS ====================
...